+++ library/src/main/java/io/sfrei/tracksearch/clients/helper/ClientHelper.java
@@ -4,22 +4,25 @@ import io.sfrei.tracksearch.clients.TrackSearchClient;
 import io.sfrei.tracksearch.clients.setup.Client;
 import io.sfrei.tracksearch.exceptions.TrackSearchException;
 import io.sfrei.tracksearch.tracks.Track;
-import lombok.extern.slf4j.Slf4j;
+import org.slf4j.Logger;
 
 import java.util.Optional;
 import java.util.function.Function;
 
-@Slf4j
-public final class ClientHelper {
 
-    public static <T extends Track> Optional<String> getStreamUrl(TrackSearchClient<T> searchClient, T track,
+public interface ClientHelper {
+
+    Logger logger();
+
+     default <T extends Track> Optional<String> getStreamUrl(TrackSearchClient<T> searchClient, T track,
                                                                   Function<String, Integer> requestForCodeFunction,
                                                                   final int retries) {
 
+        logger().trace("Get stream url for: {}", track);
         return tryToGetStreamUrl(searchClient, track, requestForCodeFunction, retries + 1);
     }
 
-    private static <T extends Track> Optional<String> tryToGetStreamUrl(TrackSearchClient<T> searchClient, T track,
+    private <T extends Track> Optional<String> tryToGetStreamUrl(TrackSearchClient<T> searchClient, T track,
                                                                         Function<String, Integer> requestForCodeFunction,
                                                                         int tries) {
 
@@ -33,12 +36,12 @@ public final class ClientHelper {
                 return Optional.ofNullable(streamUrl);
             else {
                 tries -= 1;
-                log.warn("Error getting stream URL for {} - {} retries left",
+                logger().warn("Not able getting stream URL for {} - {} retries left",
                         searchClient.getClass().getSimpleName(), tries);
                 return tryToGetStreamUrl(searchClient, track, requestForCodeFunction, tries);
             }
         } catch (TrackSearchException e) {
-            log.error("Error getting stream URL for {}", e, searchClient.getClass().getSimpleName());
+            logger().error("Error getting stream URL for {}", e, searchClient.getClass().getSimpleName());
             return Optional.empty();
         }