+++ library/src/test/java/io/sfrei/tracksearch/clients/youtube/YouTubeClientTest.java
@@ -2,19 +2,16 @@ package io.sfrei.tracksearch.clients.youtube;
 
 import io.sfrei.tracksearch.clients.ClientTestImpl;
 import io.sfrei.tracksearch.tracks.YouTubeTrack;
-import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.Tag;
 
 import java.util.Collections;
 
 import static io.sfrei.tracksearch.clients.ClientTestConstants.DEFAULT_SEARCH_KEY;
-
-@Slf4j
 @Tag("ClientTest")
 public class YouTubeClientTest extends ClientTestImpl<YouTubeTrack> {
 
     public YouTubeClientTest() {
-        super(new YouTubeClient(), Collections.singletonList(DEFAULT_SEARCH_KEY), log);
+        super(new YouTubeClient(), Collections.singletonList(DEFAULT_SEARCH_KEY));
     }
 
 }